--- a/intel_extension_for_transformers/llm/runtime/graph/core/ne_layers.c
+++ b/intel_extension_for_transformers/llm/runtime/graph/core/ne_layers.c
@@ -3126,15 +3126,16 @@ struct ne_tensor* ne_flash_attn(struct ne_context* ctx, struct ne_tensor* q, str
   NE_ASSERT(ne_can_mul_mat(k, q));
   int batch = q->ne[3];
   int headnum = q->ne[2];
+  int heads_kv = k->ne[2];
   int seq_cur = q->ne[1];
   int headsize = q->ne[0];
   int seq_all = k->ne[1];
   // int seq_past = seq_all - seq_cur;
+  NE_ASSERT(("headnum must be a multiple of heads_kv", headnum % heads_kv == 0));
   NE_ASSERT(headsize == k->ne[0]);
   NE_ASSERT(headsize == v->ne[1]);
   NE_ASSERT(seq_all == v->ne[0]);
-  NE_ASSERT(headnum == k->ne[2]);
-  NE_ASSERT(headnum == v->ne[2]);
+  NE_ASSERT(("n_heads must be the same for K/V!", k->ne[2] == v->ne[2]));
   NE_ASSERT(batch == k->ne[3]);
   NE_ASSERT(batch == v->ne[3]);
   bool is_node = true;
@@ -8821,12 +8822,12 @@ static void ne_compute_forward_flash_attn_f32_f16_f16(const struct ne_compute_pa
 
   const int64_t nek0 = k->ne[0];
   const int64_t nek1 = k->ne[1];
-  // const int64_t nek2 = k->ne[2];
+  const int64_t nek2 = k->ne[2];
   // const int64_t nek3 = k->ne[3];
 
   // const int64_t nev0 = v->ne[0];
   const int64_t nev1 = v->ne[1];
-  // const int64_t nev2 = v->ne[2];
+  const int64_t nev2 = v->ne[2];
   // const int64_t nev3 = v->ne[3];
 
   const int64_t ne0 = dst->ne[0];
@@ -8856,19 +8857,15 @@ static void ne_compute_forward_flash_attn_f32_f16_f16(const struct ne_compute_pa
 
   const int64_t headsize = neq0;
   const int64_t headnum = neq2;
+  const int64_t heads_kv = nek2;
   const int64_t embedsize = headnum * headsize;
   const int64_t seq_cur = neq1;
   const int64_t seq_all = nek1;
   const int64_t seq_past = seq_all - seq_cur;
   const int64_t batch = neq3;
 
-  if (params->type == NE_TASK_INIT) {
-    return;
-  }
+  if (params->type == NE_TASK_INIT || params->type == NE_TASK_FINALIZE) return;
 
-  if (params->type == NE_TASK_FINALIZE) {
-    return;
-  }
   const int keles = ne_element_size(k);
   const int veles = ne_element_size(v);
   int step_k_sl = k->nb[1] / keles;
@@ -8895,6 +8892,7 @@ static void ne_compute_forward_flash_attn_f32_f16_f16(const struct ne_compute_pa
       .attn_flags = flags,
       .batch_size = batch,
       .head_num = headnum,
+      .heads_kv = heads_kv,
       .head_size = headsize,
       .sl_q = seq_cur,
       .sl_kv = seq_all,
@@ -8924,9 +8922,11 @@ static void ne_compute_forward_flash_attn_reordered(const struct ne_compute_para
                                                     const struct ne_tensor* k, const struct ne_tensor* v,
                                                     const struct ne_tensor* tmp, struct ne_tensor* dst) {
   if (params->type != NE_TASK_COMPUTE) return;
+
   const int64_t headsize = q->ne[0];
   const int64_t seq_cur = q->ne[1];
   const int64_t headnum = q->ne[2];
+  const int64_t heads_kv = k->ne[2];
   const int64_t batch = q->ne[3];
   const int64_t embedsize = headnum * headsize;
   const int64_t seq_all = k->ne[1];
@@ -8953,6 +8953,7 @@ static void ne_compute_forward_flash_attn_reordered(const struct ne_compute_para
       .attn_flags = flags,
       .batch_size = batch,
       .head_num = headnum,
+      .heads_kv = heads_kv,
       .head_size = headsize,
       .sl_q = seq_cur,
       .sl_kv = seq_all,
@@ -9237,7 +9238,7 @@ static void ne_compute_forward_flash_attn_kv_update(const struct ne_compute_para
       .src = cur->data,
       .cache = cache->data,
       .batch_size = cur->ne[3],
-      .head_num = cur->ne[1],
+      .heads_kv = cur->ne[1],
       .head_size = cur->ne[0],
       .seq_off = n_past,
       .seq_size = cur->ne[2],