--- a/intel_extension_for_transformers/llm/runtime/graph/core/ne_layers.h
+++ b/intel_extension_for_transformers/llm/runtime/graph/core/ne_layers.h
@@ -79,6 +79,14 @@
 extern "C" {
 #endif
 
+// Attention flags
+typedef enum NE_ATTN_FLAG {
+  NE_ATTN_FLAG_IS_CAUSAL = 1 << 1,
+  NE_ATTN_FLAG_IS_ALIBI = 1 << 2,
+} NE_ATTN_FLAG;
+typedef uint32_t ne_attn_flags_t;
+
+
 // convert FP16 <-> FP32
 NE_API float ne_fp16_to_fp32(ne_fp16_t x);
 NE_API ne_fp16_t ne_fp32_to_fp16(float x);
@@ -407,7 +415,7 @@ NE_API struct ne_tensor* ne_conv_1d_1s(struct ne_context* ctx, struct ne_tensor*
 NE_API struct ne_tensor* ne_conv_1d_2s(struct ne_context* ctx, struct ne_tensor* a, struct ne_tensor* b);
 
 NE_API struct ne_tensor* ne_flash_attn(struct ne_context* ctx, struct ne_tensor* q, struct ne_tensor* k,
-                                       struct ne_tensor* v, float scale, bool masked);
+                                       struct ne_tensor* v, float scale, ne_attn_flags_t flags);
 NE_API struct ne_tensor* ne_flash_attn_update_k(struct ne_context* ctx, struct ne_tensor* cache, struct ne_tensor* cur,
                                                 int n_past);
 NE_API struct ne_tensor* ne_flash_attn_update_v(struct ne_context* ctx, struct ne_tensor* cache, struct ne_tensor* cur,