@DDEle 2023-10-09
BTW, @a32543254 I added a flexible OP for debug experience. It allows developers to access forwarding result in the graph file (with help of non-capture lambda expression).

https://github.com/intel/intel-extension-for-transformers/commit/69ee8ebce4239646d4f1126ef2d9b3bfe3934d87#diff-fd662f3dbbde39aca9f4a1df0017d713ee7c5aa508025b36315c6defac152bc0R178-R184